+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Update `Generate context output` to `Devicetree Context Output`
+- Fixed missing compatible in serialized output for node with `simple-bus` type.
 
 ## [0.10.0] - 2026-05-28